change from sqlAPI call kafka producer to kafka consumer call sqlAPI
